start reworking and organizing the code base
