+++ src/lib.rs
@@ -19,42 +19,13 @@ use crossterm::QueueableCommand;
 use crate::adapted_input::{AdaptedKeyboardInput, RawConsoleEvent};
 
 mod adapted_input;
-
-/// By default the loop will target 4 FPS
-const DEFAULT_LOOP_DELAY: Duration = Duration::from_millis(2500);
+mod scheduler;
+mod terminal_helpers;
 
 #[derive(Resource)]
 pub struct Terminal<T: tui::backend::Backend>(pub tui::Terminal<T>);
 
 pub type BevyTerminal = Terminal<tui::backend::CrosstermBackend<std::io::Stdout>>;
-
-#[derive(Resource)]
-struct TuiPersistentState {
-    first_run: bool,
-    last_update: Instant,
-    timeout_reached: bool,
-}
-
-impl TuiPersistentState {
-    fn is_first_run(&self) -> bool {
-        self.first_run
-    }
-
-    fn mark_completed_tick(&mut self) {
-        self.first_run = false;
-        self.last_update = Instant::now();
-    }
-}
-
-impl Default for TuiPersistentState {
-    fn default() -> Self {
-        Self {
-            first_run: true,
-            last_update: Instant::now(),
-            timeout_reached: false,
-        }
-    }
-}
 
 pub struct MinimalTuiPlugins;
 
@@ -92,124 +63,7 @@ impl Plugin for TuiPlugin {
     }
 }
 
-fn event_handler(app: &mut App, event: Event) {
-    match event {
-        Event::FocusGained => {
-            // todo: handle marking us as actively focused in our window equivalent
-        },
-        Event::FocusLost => {
-            // todo: handle marking us as no longer focused in our window equivalent
-        },
-        Event::Key(event) => {
-            adapted_input::convert_adapted_keyboard_input(&event)
-                .into_iter()
-                .for_each(|ki| app.world.send_event(ki));
-        }
-        Event::Mouse(_event) => {
-            // todo: begin handling mouse events
-        }
-        Event::Paste(ref _data) => {
-            // todo: publish event with the pasted content
-            // todo: do I get style info?
-        }
-        Event::Resize(_width, _height) => {
-            // todo: update the size of our window equivalent
-        }
-    }
-
-    app.world.send_event(RawConsoleEvent(event));
-}
-
-pub fn create_terminal() -> Result<BevyTerminal, Box<dyn std::error::Error>> {
-    let stdout = std::io::stdout();
-    let backend = tui::backend::CrosstermBackend::new(stdout);
-    let terminal = tui::Terminal::new(backend)?;
-    Ok(Terminal(terminal))
-}
-
-pub fn initialize_terminal() -> Result<(), Box<dyn std::error::Error>> {
-    crossterm::terminal::enable_raw_mode()?;
-
-    let mut stdout = std::io::stdout();
-    stdout.queue(crossterm::terminal::EnterAlternateScreen)?;
-    stdout.queue(crossterm::event::EnableBracketedPaste)?;
-    stdout.queue(crossterm::event::EnableFocusChange)?;
-    stdout.queue(crossterm::event::EnableMouseCapture)?;
-    //stdout.queue(crossterm::event::PushKeyboardEnhancementFlags(
-    //    crossterm::event::KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES
-    //        | crossterm::event::KeyboardEnhancementFlags::REPORT_EVENT_TYPES
-    //        | crossterm::event::KeyboardEnhancementFlags::REPORT_ALL_KEYS_AS_ESCAPE_CODES,
-    //))?;
-    stdout.flush().expect("terminal command trigger");
-
-    Ok(())
-}
-
-pub fn teardown_terminal() -> Result<(), Box<dyn std::error::Error>> {
-    crossterm::terminal::disable_raw_mode()?;
-
-    let mut stdout = std::io::stdout();
-    stdout.queue(crossterm::terminal::LeaveAlternateScreen)?;
-    stdout.queue(crossterm::event::DisableBracketedPaste)?;
-    stdout.queue(crossterm::event::DisableFocusChange)?;
-    stdout.queue(crossterm::event::DisableMouseCapture)?;
-    //stdout.queue(crossterm::event::PopKeyboardEnhancementFlags)?;
-    stdout.queue(crossterm::cursor::Show)?;
-    stdout.flush()?;
-
-    Ok(())
-}
-
 fn terminal_setup(mut commands: Commands) {
     let term = create_terminal().expect("terminal setup to succeed");
     commands.insert_resource(term);
-}
-
-fn tick(
-    app: &mut App,
-    app_exit_event_reader: &mut ManualEventReader<AppExit>,
-) -> Result<Option<Duration>, Box<dyn std::error::Error>> {
-    let start_time = Instant::now();
-
-    // The app needs to tick once to allow the startup system to setup the terminal. We delay any
-    // event processing until this is available otherwise this would become a blocking call until
-    // an event is received.
-    let first_run = app.world.resource::<TuiPersistentState>().is_first_run();
-    if !first_run {
-        // todo: need to adjust this delay based on how long the last loop took
-        let events_available = poll_term(DEFAULT_LOOP_DELAY)?;
-
-        if events_available {
-            // Read all of the available events all at once
-            while poll_term(Duration::from_secs(0))? {
-                event_handler(app, read_term()?);
-            }
-        }
-
-        // Indicate that this tick was triggered by the timeout and not by an event
-        app.world
-            .resource_mut::<TuiPersistentState>()
-            .timeout_reached = !events_available;
-    }
-
-    app.update();
-    app.world
-        .resource_mut::<TuiPersistentState>()
-        .mark_completed_tick();
-
-    if let Some(app_exit_events) = app.world.get_resource::<Events<AppExit>>() {
-        if app_exit_event_reader.iter(app_exit_events).last().is_some() {
-            return Ok(None);
-        }
-    }
-
-    Ok(Some(Instant::now() - start_time))
-}
-
-fn tui_schedule_runner(mut app: App) {
-    let mut app_exit_event_reader = ManualEventReader::<AppExit>::default();
-
-    while let Ok(Some(_tick_duration)) = tick(&mut app, &mut app_exit_event_reader) {
-        // more stuff to do
-    }
 }
